requirements.txt
streamlit
pillow
numpy
python-docx